Refactor to get login/logout to work
