+++ src/InMemoryLists.js
@@ -5,6 +5,12 @@ import firebase from "firebase/compat";
 import {useCollection} from "react-firebase-hooks/firestore";
 import {useState} from "react";
 import Lists from "./Lists";
+import TabList from './TabList';
+import {
+    useAuthState,
+    useCreateUserWithEmailAndPassword,
+    useSignInWithEmailAndPassword
+} from 'react-firebase-hooks/auth';
 
 const firebaseConfig = {
     apiKey: "AIzaSyDGHngb4moL3bpSKohoDA3nUtkfC8YMwp4",
@@ -27,6 +33,83 @@ const auth = firebase.auth();
 const googleProvider = new firebase.auth.GoogleAuthProvider();
 
 function InMemoryLists() {
+    const [user, loading, error] = useAuthState(auth);
+
+    function verifyEmail() {
+        auth.currentUser.sendEmailVerification();
+    }
+
+    if (loading) {
+        return <p>Checking...</p>;
+    } else if (user) {
+        return <div>
+            {user.displayName || user.email}
+            <SignedInApp user={user}/>
+            <button type="button" onClick={() => auth.signOut()}>Logout</button>
+            {!user.emailVerified && <button type="button" onClick={verifyEmail}>Verify email</button>}
+        </div>
+    } else {
+        return <>
+            {error && <p>Error App: {error.message}</p>}
+            <TabList>
+                <SignIn key="Sign In" />
+                <SignUp key="Sign Up" />
+            </TabList>
+        </>
+    }
+}
+
+const FAKE_EMAIL = 'foo@bar.com';
+const FAKE_PASSWORD = 'xyzzyxx';
+
+function SignIn() {
+    const [
+        signInWithEmailAndPassword,
+        userCredential, loading, error
+    ] = useSignInWithEmailAndPassword(auth);
+
+    if (userCredential) {
+        // Shouldn't happen because App should see that
+        // we are signed in.
+        return <div>Unexpectedly signed in already</div>
+    } else if (loading) {
+        return <p>Logging in…</p>
+    }
+    return <div>
+        {error && <p>"Error logging in: " {error.message}</p>}
+        <button onClick={() =>
+            signInWithEmailAndPassword(FAKE_EMAIL, FAKE_PASSWORD)}>Login with test user Email/PW
+        </button>
+        <button onClick={() =>
+            auth.signInWithPopup(googleProvider)}>Login with Google
+        </button>
+    </div>
+}
+
+function SignUp() {
+    const [
+        createUserWithEmailAndPassword,
+        userCredential, loading, error
+    ] = useCreateUserWithEmailAndPassword(auth);
+
+    if (userCredential) {
+        // Shouldn't happen because App should see that
+        // we are signed in.
+        return <div>Unexpectedly signed in already</div>
+    } else if (loading) {
+        return <p>Signing up…</p>
+    }
+    return <div>
+        {error && <p>"Error signing up: " {error.message}</p>}
+        <button onClick={() =>
+            createUserWithEmailAndPassword(FAKE_EMAIL, FAKE_PASSWORD)}>
+            Create test user
+        </button>
+
+    </div>
+}
+
+function SignedInApp() {
     const [currentListId, setCurrentListId] = useState("");
     const [currentListName, setCurrentListName] = useState("");
     const query = db.collection(collectionName).orderBy("id");
